Move index to bip32.ts, makes exported methods easier to read
